separate ComposedRegions test from memory
