+++ Tests/InterpreterTests/MemoryTests.swift
@@ -30,130 +30,6 @@ final class InterpreterMemoryTests: XCTestCase {
     }
   }
 
-  func testTupleComposeDecompose() throws {
-    var l = TypeLayoutCache(typesIn: TypedProgram.empty, for: UnrealABI())
-    let i16Pair = ^TupleType(types: [.builtin(.i(16)), .builtin(.i(16))])
-    let i16 = ^BuiltinType.i(16)
-
-    assert(l[i16Pair].alignment > 1, "Need to produce misaligned access for testing")
-
-    var m = Memory(typesIn: TypedProgram.empty, for: UnrealABI())
-    let p = m.allocate(i16Pair).address
-
-    check(throws: Memory.Error.alignment(p + 1, for: l[i16])) {
-      try m.compose(i16, at: p + 1)
-    }
-
-    // An address that would be suitably aligned, but out of bounds for
-    // i16Pair initialization.  The initialized object would extend past
-    // the end of the allocation.
-    let outOfBoundsFori16Pair = p + l[i16].size
-    check(
-      throws: Memory.Error.bounds(outOfBoundsFori16Pair, for: l[i16Pair], allocationSize: l[i16Pair].size)
-    ) {
-      try m.compose(i16Pair, at: outOfBoundsFori16Pair)
-    }
-
-    let parts = l[i16Pair].parts
-    let partIDs = Array(l[i16Pair].partParentages)
-    check(throws: Memory.Error.noComposedPart(at: p, partIDs[0])) {
-      try m.compose(i16Pair, at: p)
-    }
-
-    try m.compose(i16, at: p + parts[0].offset)
-
-    check(throws: Memory.Error.noComposedPart(at: p + parts[1].offset, partIDs[1])) {
-      try m.compose(i16Pair, at: p)
-    }
-
-    try m.compose(i16, at: p + parts[1].offset)
-
-    try m.compose(i16Pair, at: p)
-
-    check(throws: Memory.Error.noDecomposable(l[i16], at: p)) {
-      try m.decompose(i16, at: p)
-    }
-
-    try m.decompose(i16Pair, at: p)
-
-    check(throws:Memory.Error.noDecomposable(l[i16Pair], at: p)) {
-      try m.decompose(i16Pair, at: p)
-    }
-
-    try m.decompose(i16, at: p + parts[0].offset)
-
-    check(throws: Memory.Error.noDecomposable(l[i16], at: p + parts[0].offset)) {
-      try m.decompose(i16, at: p)
-    }
-
-    try m.decompose(i16, at: p + parts[1].offset)
-  }
-
-  func testUnionComposeDecompose() throws {
-    var l = TypeLayoutCache(typesIn: TypedProgram.empty, for: UnrealABI())
-    let i16 = ^BuiltinType.i(16)
-    let i32 = ^BuiltinType.i(32)
-    let i16i32Union = ^UnionType([i16, i32])
-
-    var m = Memory(typesIn: TypedProgram.empty, for: UnrealABI())
-    let p = m.allocate(i16i32Union).address
-
-    assert(l[i16i32Union].alignment > 1)
-
-    check(throws: Memory.Error.alignment(p + 1, for: l[i16i32Union])) {
-      try m.compose(i16i32Union, at: p + 1)
-    }
-
-    // An address that would be suitably aligned, but out of bounds for
-    // i16i32Union initialization.  The initialized object would extend past
-    // the end of the allocation.
-    let outOfBoundsFori16i32Union = p + l[i32].size
-    check(
-      throws: Memory.Error.bounds(outOfBoundsFori16i32Union, for: l[i16i32Union], allocationSize: l[i16i32Union].size)
-    ) {
-      try m.compose(i16i32Union, at: outOfBoundsFori16i32Union)
-    }
-
-    let parts = l[i16i32Union].parts
-    let partIDs = Array(l[i16i32Union].partParentages)
-    let discriminator = parts.last!
-    check(throws: Memory.Error.noComposedPart(at: p + discriminator.offset, partIDs.last!)) {
-      try m.compose(i16i32Union, at: p)
-    }
-    _ = i16
-    /*
-      m.compose()
-      try m.compose(i16, at: p + parts[0].offset)
-
-      XCTAssertThrowsError(Memory.Error.noComposedPart(at: p + parts[1].offset, partIDs[1])) {
-        try m.compose(i16i32Union, at: p)
-      }
-
-      try m.compose(i16, at: p + parts[1].offset)
-
-      try m.compose(i16i32Union, at: p)
-
-      XCTAssertThrowsError(Memory.Error.noDecomposable(i16, at: p)) {
-        try m.decompose(i16, at: p)
-      }
-
-      try m.decompose(i16i32Union, at: p)
-
-      XCTAssertThrowsError(Memory.Error.noDecomposable(i16i32Union, at: p)) {
-        try m.decompose(i16i32Union, at: p)
-      }
-
-      try m.decompose(i16, at: p + parts[0].offset)
-
-      XCTAssertThrowsError(Memory.Error.noDecomposable(i16, at: p + parts[0].offset)) {
-        try m.decompose(i16, at: p)
-      }
-
-      try m.decompose(i16, at: p + parts[1].offset)
-     */
-  }
-
-  
   func testSubPartLayout() throws {
     var m = Memory(typesIn: TypedProgram.empty, for: UnrealABI())
     let i8 = ^BuiltinType.i(8);
@@ -177,26 +53,28 @@ final class InterpreterMemoryTests: XCTestCase {
     // TODO: add test for union case.
   }
 
-  func testTypeBindUnbind() throws {
-    var l = TypeLayoutCache(typesIn: TypedProgram.empty, for: UnrealABI())
-    var m = Memory(typesIn: TypedProgram.empty, for: UnrealABI())
-    let i64 = ^BuiltinType.i(64)
-    let tuple = ^TupleType(types: [i64, i64])
-    let p = m.allocate(tuple, count: 2)
-
-    check(throws: Memory.Error.alignment(p.address + 1, for: l[i64])) {
-      try m.bind(type: i64, at: p.address + 1)
-    }
-    check(throws: Memory.Error.bounds(p.address + 32, for: l[tuple], allocationSize: 32)) {
-      try m.bind(type: tuple, at: p.address + 32)
-    }
-
-    try m.bind(type: tuple, at: p.address)
-    check(throws: Memory.Error.regionAlreadyReserved(for: tuple)) {
-      try m.bind(type: i64, at: p.address)
-    }
-    m.unbindType(from: p.address)
-    try m.bind(type: tuple, at: p.address)
-    try m.bind(type: tuple, at: p.address + 16)
-  }
+  // TODO: port it to new component
+  //
+  // func testTypeBindUnbind() throws {
+  //   var l = TypeLayoutCache(typesIn: TypedProgram.empty, for: UnrealABI())
+  //   var m = Memory(typesIn: TypedProgram.empty, for: UnrealABI())
+  //   let i64 = ^BuiltinType.i(64)
+  //   let tuple = ^TupleType(types: [i64, i64])
+  //   let p = m.allocate(tuple, count: 2)
+  //
+  //   check(throws: Memory.Error.alignment(p.address + 1, for: l[i64])) {
+  //     try m.bind(type: i64, at: p.address + 1)
+  //   }
+  //   check(throws: Memory.Error.bounds(p.address + 32, for: l[tuple], allocationSize: 32)) {
+  //     try m.bind(type: tuple, at: p.address + 32)
+  //   }
+  //
+  //   try m.bind(type: tuple, at: p.address)
+  //   check(throws: Memory.Error.regionAlreadyReserved(for: tuple)) {
+  //     try m.bind(type: i64, at: p.address)
+  //   }
+  //   m.unbindType(from: p.address)
+  //   try m.bind(type: tuple, at: p.address)
+  //   try m.bind(type: tuple, at: p.address + 16)
+  // }
 }
